+++ inboxen/_vendor/cursor_pagination.py
@@ -1,5 +1,5 @@
 from base64 import b64decode, b64encode
-from collections import Sequence
+from collections.abc import Sequence
 
 from django.db.models import Field, Func, Value, TextField
 from django.utils.translation import gettext_lazy as _