+++ inboxen/tickets/tests.py
@@ -338,7 +338,7 @@ class QuestionAdminIndexTestCase(InboxenTestCase):
         self.user = factories.UserFactory(is_superuser=True)
 
     def test_url(self):
-        assert self.client.login(username=self.user.username, password="123456", request=MockRequest(self.user)),\
+        assert self.client.login(username=self.user.username, password="123456", request=MockRequest(self.user)), \
             "Could not log in"
 
         grant_otp(self.client, self.user)
@@ -368,7 +368,7 @@ class QuestionAdminResponseTestCase(InboxenTestCase):
         self.user = factories.UserFactory(is_superuser=True)
 
     def test_url(self):
-        assert self.client.login(username=self.user.username, password="123456", request=MockRequest(self.user)),\
+        assert self.client.login(username=self.user.username, password="123456", request=MockRequest(self.user)), \
                 "Could not log in"
 
         grant_otp(self.client, self.user)